v0.32 fixed more errors

+++ main.cpp
@@ -12,7 +12,7 @@ int main() {
 
 
 
-    power_in<float> usb3 ("usb_c", 5, 1.2);
+    power_in<double> usb3 ("usb_c", 5, 1.2);
     cout<<usb3<<"\n";
     power_in<int> usb2 ("usb_c", 5, 1);
     cout<<usb2;
